+++ Mcode/+hello/globals.m
@@ -1,0 +1,44 @@
+classdef globals
+  % Global library properties and settings for matlab-hello-world.
+  %
+  % Note that if you want to change the settings, you can't do this:
+  %
+  %    hello.globals.settings.someSetting = 42;
+  %
+  % That will break due to how Matlab Constant properties work. Instead, you need
+  % to first grab the Settings object and store it in a variable, and then work
+  % on that:
+  %
+  %    s = hello.globals.settings;
+  %    s.someSetting = 42;
+  
+  properties (Constant)
+    % Path to the root directory of this matlab-hello-world distribution
+    distroot = string(fileparts(fileparts(fileparts(mfilename('fullpath')))));
+    % Global settings for hello.
+    settings = hello.Settings.discover
+  end
+  
+  methods (Static)
+    
+    function out = version
+      % The version of the matlab-hello-world library
+      %
+      % Returns a string.
+      persistent val
+      if isempty(val)
+        versionFile = fullfile(hello.globals.distroot, 'VERSION');
+        val = strtrim(hello.internal.util.readtext(versionFile));
+      end
+      out = val;
+    end
+    
+    function initialize
+      % Initialize this library/package
+      hello.internal.initializePackage;
+    end
+    
+  end
+  
+end
+