+++ dev-kit/private/reporoot.m
@@ -1,4 +1,4 @@
 function out = reporoot
-% The root dir of the myproject repo
+% The root dir of the matlab-hello-world repo
 out = string(fileparts(fileparts(fileparts(mfilename('fullpath')))));
 end