+++ dev-kit/private/rmrf.m
@@ -1,4 +1,4 @@
 function rmrf(files)
 % Recursively delete files and folders
-mypackage.internal.util.rmrf(files);
+hello.internal.util.rmrf(files);
 end