+++ dev-kit/util-shims/fclose2.m
@@ -1,4 +1,4 @@
 function fclose2(fid)
 % A version of fclose that errors on failure
-mypackage.internal.util.fclose2(fid);
+hello.internal.util.fclose2(fid);
 end