+++ dev-kit/util-shims/frewind2.m
@@ -1,4 +1,4 @@
 function frewind2(fid)
 % A version of frewind that errors on failure
-mypackage.internal.util.frewind2(fid);
+hello.internal.util.frewind2(fid);
 end