+++ dev-kit/util-shims/rmdir2.m
@@ -1,4 +1,4 @@
 function rmdir2(dir, varargin)
 % A version of rmdir that raises errors on failure
-mypackage.internal.util.rmdir2(dir, varargin{:});
+hello.internal.util.rmdir2(dir, varargin{:});
 end