+++ CHANGES.md
@@ -1,4 +1,4 @@
-__myproject__ Changelog
+matlab-hello-world Changelog
 ================================
 
 Version 0.1.0 (date goes here)